requirements.txt
beautifulsoup4
streamlit
yfinance
pandas
requests
expiringdict
llama-cpp-python
huggingface-hub